+++ legacy-ecommerce/ecommerce-service/src/main/java/com/legacy/shop/ecommerce/dto/CartResponse.java
@@ -1,11 +1,12 @@
 package com.legacy.shop.ecommerce.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record CartResponse(
         Long id,
         Long customerId,
         List<CartItemResponse> items,
-        double total
+        BigDecimal total
 ) {
 }
