+++ legacy-ecommerce/ecommerce-service/src/main/java/com/legacy/shop/ecommerce/dto/CreateProductRequest.java
@@ -1,8 +1,10 @@
 package com.legacy.shop.ecommerce.dto;
 
+import java.math.BigDecimal;
+
 public record CreateProductRequest(
         String name,
-        double price,
+        BigDecimal price,
         Long categoryId,
         String description,
         int initialStock
